pyproject.toml
[tool.poetry]
name = "kanonymity"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
pandas = "^1.5.0"
tqdm = "^4.64.1"
prettytable = "^3.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
